@@ -6,24 +6,31 @@
   "dependencies": {
     "@babel/core": "^7.4.5",
     "@babel/plugin-proposal-class-properties": "^7.4.4",
+    "@babel/plugin-proposal-export-namespace-from": "^7.2.0",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "@babel/plugin-transform-runtime": "^7.4.4",
     "@babel/preset-env": "^7.4.5",
     "@hot-loader/react-dom": "^16.8.6",
-    "babel-eslint": "^10.0.1",
+    "babel-eslint": "^10.0.2",
     "babel-loader": "^8.0.6",
     "babel-plugin-lodash": "^3.3.4",
     "babel-plugin-module-resolver": "^3.2.0",
+    "big-integer": "^1.6.44",
     "classnames": "^2.2.6",
     "css-loader": "^3.0.0",
-    "eslint": "^5.16.0",
+    "eslint": "^6.0.1",
+    "eslint-plugin-import": "^2.18.0",
     "extract-loader": "^3.1.0",
     "file-loader": "^4.0.0",
-    "firebase-tools": "^7.0.0",
+    "firebase-tools": "^7.0.2",
     "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^3.2.0",
     "immutable": "^4.0.0-rc.12",
+    "json5": "^2.1.0",
     "lodash": "^4.17.11",
     "lodash-webpack-plugin": "^0.11.5",
+    "mini-css-extract-plugin": "^0.7.0",
+    "nodemon": "^1.19.1",
     "normalize.css": "^8.0.1",
     "nps": "^5.9.5",
     "nps-utils": "^1.7.0",
@@ -38,7 +45,7 @@
     "postcss-url": "^8.0.0",
     "precss": "^4.0.0",
     "prop-types": "^15.7.2",
-    "pug": "^2.0.3",
+    "pug": "^2.0.4",
     "pug-extract-loader": "^1.0.8",
     "pug-loader": "^2.4.0",
     "pug-plain-loader": "^1.0.0",
@@ -50,22 +57,18 @@
     "redux": "^4.0.1",
     "redux-immutable": "^4.0.0",
     "reselect": "^4.0.0",
+    "script-ext-html-webpack-plugin": "^2.1.3",
     "stylelint": "^10.1.0",
     "stylelint-config-standard": "^18.3.0",
     "sugarss": "^2.0.0",
     "terser": "^4.0.0",
     "terser-webpack-plugin": "^1.3.0",
-    "webpack": "^4.34.0",
+    "webpack": "^4.35.0",
     "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-cli": "^3.3.4",
-    "webpack-dev-server": "^3.7.1",
+    "webpack-cli": "^3.3.5",
+    "webpack-dev-server": "^3.7.2",
     "webpack-merge": "^4.2.1",
     "webpack-stylish": "^0.1.8"
-  },
-  "devDependencies": {
-    "pug-html-loader": "^1.1.5",
-    "restyle-loader": "^0.0.3",
-    "style-loader": "^0.23.1"
   },
   "license": "MIT"
 }
