@@ -1,28 +1,30 @@
 
-const p__lodash = require('lodash/throttle')
+import * as p__json5 from 'json5'
+import * as p__lodash from 'lodash'
 
-const m__token = require('~/redux/token')
-const m__windows = require('~/commons/windows')
+import * as m__token from '~/redux/token'
+import * as m__windows from '~/commons/windows'
+
 
 const storage = m__windows.local_storage
 
-exports.deserialize = () => {
+export const deserialize = () => {
 	try {
 		const string = storage.getItem(m__token.value)
 		if (string === null) {
 			return undefined
 		}
-		return JSON.parse(string)
+		return p__json5.parse(string)
 	} catch (error) {
 		console.warn(error)
 	}
 }
 
-exports.store_serializer = (store) => {
+export const store_serializer = (store) => {
 	return () => {
 		const state = store.getState()
 		try {
-			const string = JSON.stringify(state)
+			const string = p__json5.stringify(state)
 			storage.setItem(m__token.value, string)
 		} catch (error) {
 			console.warn(error)
@@ -30,8 +32,8 @@ exports.store_serializer = (store) => {
 	}
 }
 
-exports.store_throttler = (store, milliseconds = 1 << 12) => {
-	return p__lodash(exports.store_serializer(store), milliseconds, {
+export const store_throttler = (store, milliseconds = 1 << 12) => {
+	return p__lodash.throttle(store_serializer(store), milliseconds, {
 		leading: false,
 	})
 }
