@@ -1,17 +1,20 @@
 
-const p__immutable = require('immutable')
-const p__react = require('react')
-const p__react_hot_loader__root = require('react-hot-loader/root')
-const p__react_hyperscript = require('react-hyperscript')
-const p__react_redux = require('react-redux')
+import * as p__immutable from 'immutable'
+import * as p__react_redux from 'react-redux'
 
-const m__container__kanban_board = require('~/containers/kanban-board')
-const m__local_storage = require('~/commons/local-storage')
-const m__redux_store = require('~/commons/redux-store')
+import i__react from 'react'
+import i__react_hot_loader__root from 'react-hot-loader/root'
+import i__react_hyperscript from 'react-hyperscript'
 
-const $ = p__react_hyperscript
+import * as m__local_storage from '~/commons/local-storage'
 
-const component = class extends p__react.PureComponent {
+import d__container__kanban_board from '~/containers/kanban-board'
+import d__redux_store from '~/commons/redux-store'
+
+
+const $ = i__react_hyperscript
+
+const component = class extends i__react.PureComponent {
 	static displayName = 'app'
 	componentWillMount() {
 		const state = m__local_storage.deserialize()
@@ -20,12 +23,12 @@ const component = class extends p__react.PureComponent {
 	render() {
 		return [
 			$(p__react_redux.Provider, {
-				store: m__redux_store(this.immutable),
+				store: d__redux_store(this.immutable),
 			}, [
-				$(m__container__kanban_board),
+				$(d__container__kanban_board),
 			]),
 		][0]
 	}
 }
 
-module.exports = p__react_hot_loader__root.hot(component)
+export default i__react_hot_loader__root.hot(component)
