@@ -1,29 +1,37 @@
 
-const p__classnames = require('classnames')
-const p__immutable = require('immutable')
-const p__prop_types = require('prop-types')
-const p__react = require('react')
-const p__react_hyperscript = require('react-hyperscript')
+import * as p__immutable from 'immutable'
 
-const m__container__task = require('../../containers/task')
+import i__classnames from 'classnames/bind'
+import i__prop_types from 'prop-types'
+import i__react from 'react'
+import i__react_hyperscript from 'react-hyperscript'
 
-const $ = p__react_hyperscript
+import * as m__helpers from '~/commons/helpers'
 
-module.exports = class extends p__react.PureComponent {
+import d__container__task from '../../containers/task'
+
+import s__styles from './styles.sss'
+
+
+const $ = i__react_hyperscript
+const classnames = i__classnames.bind(s__styles)
+const style = m__helpers.styler(s__styles)
+
+export default class extends i__react.PureComponent {
 	static displayName = 'column'
 	static propTypes = {
-		id: p__prop_types.number.isRequired,
-		name: p__prop_types.string.isRequired,
-		sorting: p__prop_types.oneOf([
+		id: i__prop_types.number.isRequired,
+		name: i__prop_types.string.isRequired,
+		sorting: i__prop_types.oneOf([
 			-1, 0, 1,
 		]).isRequired,
-		tasks: p__prop_types.objectOf(
+		tasks: i__prop_types.objectOf(
 			p__immutable.List,
 		).isRequired,
-		actions: p__prop_types.shape({
-			add: p__prop_types.func.isRequired,
-			rename: p__prop_types.func.isRequired,
-			sort: p__prop_types.func.isRequired,
+		actions: i__prop_types.shape({
+			add: i__prop_types.func.isRequired,
+			rename: i__prop_types.func.isRequired,
+			sort: i__prop_types.func.isRequired,
 		}).isRequired,
 	}
 	self = {
@@ -40,7 +48,7 @@ module.exports = class extends p__react.PureComponent {
 			id,
 		}) => {
 			return [
-				$(m__container__task, {
+				$(d__container__task, {
 					key: id,
 					id,
 				}),
@@ -49,35 +57,35 @@ module.exports = class extends p__react.PureComponent {
 	}
 	render() {
 		return [
-			$('div.column', [
+			$('div' + style('column'), [
 				$('div', {
-					className: p__classnames({
+					className: classnames({
 						'column--head': !this.props.tasks.isEmpty(),
 						'column--head--empty': this.props.tasks.isEmpty(),
 					}),
 				}, [
-					$('div.column--title', [
-						$('input.column--input--name', {
+					$('div' + style('column--title'), [
+						$('input' + style('column--input--name'), {
 							onChange: this.self.handle_change_name,
 							placeholder: 'column name',
 							value: this.props.name,
 						}),
 					]),
-					$('div.column--options', [
-						$('div.column--action--add', [
-							$('span.column--action--add--text', {
+					$('div' + style('column--options'), [
+						$('div' + style('column--action--add'), [
+							$('span' + style('column--action--add--text'), {
 								onClick: this.props.actions.add,
 							}, [
 								'add task',
 							]),
 						]),
-						$('div.column--action--sort', [
-							$('span.column--action--sort--arrow', [
+						$('div' + style('column--action--sort'), [
+							$('span' + style('column--action--sort--arrow'), [
 								!!this.props.sorting && (
 									this.props.sorting < 0 ? '↑' : '↓'
 								), ' ',
 							]),
-							$('span.column--action--sort--title', {
+							$('span' + style('column--action--sort--title'), {
 								onClick: this.props.actions.sort,
 							}, [
 								'sort',
@@ -85,7 +93,7 @@ module.exports = class extends p__react.PureComponent {
 						]),
 					]),
 				]),
-				this.props.tasks.isEmpty() || $('div.column--body', [
+				this.props.tasks.isEmpty() || $('div' + style('column--body'), [
 					this.props.tasks.map(this.self.render_task),
 				]),
 			]),
