@@ -1,24 +1,30 @@
 
-const p__classnames = require('classnames')
-const p__immutable = require('immutable')
-const p__prop_types = require('prop-types')
-const p__react = require('react')
-const p__react_hyperscript = require('react-hyperscript')
+import * as p__immutable from 'immutable'
 
-const $ = p__react_hyperscript
+import i__prop_types from 'prop-types'
+import i__react from 'react'
+import i__react_hyperscript from 'react-hyperscript'
 
-module.exports = class extends p__react.PureComponent {
+import * as m__helpers from '~/commons/helpers'
+
+import s__styles from './styles.sss'
+
+
+const $ = i__react_hyperscript
+const style = m__helpers.styler(s__styles)
+
+export default class extends i__react.PureComponent {
 	static displayName = 'task'
 	static propTypes = {
-		name: p__prop_types.string.isRequired,
-		text: p__prop_types.string.isRequired,
-		columns: p__prop_types.objectOf(
+		name: i__prop_types.string.isRequired,
+		text: i__prop_types.string.isRequired,
+		columns: i__prop_types.objectOf(
 			p__immutable.List,
 		).isRequired,
-		actions: p__prop_types.shape({
-			delete: p__prop_types.func.isRequired,
-			locate: p__prop_types.func.isRequired,
-			update: p__prop_types.func.isRequired,
+		actions: i__prop_types.shape({
+			remove: i__prop_types.func.isRequired,
+			locate: i__prop_types.func.isRequired,
+			update: i__prop_types.func.isRequired,
 		}).isRequired,
 	}
 	state = {
@@ -26,7 +32,7 @@ module.exports = class extends p__react.PureComponent {
 	}
 	constructor(props) {
 		super(props)
-		this.self.references.move = p__react.createRef()
+		this.self.references.move = i__react.createRef()
 	}
 	componentDidMount() {
 		document.addEventListener('click', this.self.click)
@@ -73,20 +79,20 @@ module.exports = class extends p__react.PureComponent {
 			name,
 		}) => {
 			return [
-				$('li.task--action--locate', {
+				$('li' + style('task--action--locate'), {
 					key: id,
 				}, [
-					$('div.task--action--locate--arrow', [
+					$('div' + style('task--action--locate--arrow'), [
 						'→ ',
 					]),
-					$('div.task--action--locate--title', {
+					$('div' + style('task--action--locate--title'), {
 						onClick: () => {
 							this.props.actions.locate({
 								column_id: id,
 							})
 						},
 					}, [
-						$('span.task--action--locate--text', [
+						$('span' + style('task--action--locate--text'), [
 							name,
 						]),
 					]),
@@ -96,47 +102,47 @@ module.exports = class extends p__react.PureComponent {
 	}
 	render() {
 		return [
-			$('div.task', [
-				$('div.task--name', [
-					$('input.task--input--name', {
+			$('div' + style('task'), [
+				$('div' + style('task--name'), [
+					$('input' + style('task--input--name'), {
 						onChange: this.self.handle_change_name,
 						placeholder: 'task name',
 						value: this.props.name,
 					}),
 				]),
-				$('div.task--text', [
-					$('input.task--input--text', {
+				$('div' + style('task--text'), [
+					$('input' + style('task--input--text'), {
 						onChange: this.self.handle_change_text,
 						placeholder: 'task text',
 						value: this.props.text,
 					}),
 				]),
-				$('div.task--options', [
-					$('div.task--action--delete', [
-						$('span.task--action--delete--text', {
-							onClick: this.props.actions.delete,
+				$('div' + style('task--options'), [
+					$('div' + style('task--action--remove'), [
+						$('span' + style('task--action--remove--text'), {
+							onClick: this.props.actions.remove,
 						}, [
-							'delete',
+							'remove',
 						]),
 					]),
-					$('div.task--dropdown', {
+					$('div' + style('task--dropdown'), {
 						ref: this.self.references.move,
 					}, [
-						$('div.task--action--move', [
-							$('span.task--action--move--text', [
+						$('div' + style('task--action--move'), [
+							$('span' + style('task--action--move--text'), [
 								this.state.moving ? $([
-									$('span.task--action--move--arrow', [
+									$('span' + style('task--action--move--arrow'), [
 										'↓',
 									]), ' ',
 								]) : '↓ ',
-								$('span.task--action--move--title', {
+								$('span' + style('task--action--move--title'), {
 									onClick: this.self.move,
 								}, [
 									'move',
 								]),
 							]),
 						]),
-						this.state.moving && $('ul.task--list', [
+						this.state.moving && $('ul' + style('task--list'), [
 							this.props.columns.map(this.self.render_column),
 						]),
 					]),
