@@ -1,12 +1,14 @@
 
-const p__react_redux = require('react-redux')
-const p__redux = require('redux')
+import * as p__react_redux from 'react-redux'
+import * as p__redux from 'redux'
 
-const m__actions = require('~/redux/actions')
-const m__component = require('./component')
-const m__selectors = require('./selectors')
+import * as m__actions from '~/redux/actions'
+import * as m__selectors from './selectors'
 
-module.exports = p__react_redux.connect(
+import d__component from './component'
+
+
+export default p__react_redux.connect(
 	(state) => {
 		return m__selectors.select(state)
 	},
@@ -15,4 +17,4 @@ module.exports = p__react_redux.connect(
 			actions: p__redux.bindActionCreators(m__actions, dispatch),
 		}
 	},
-)(m__component)
+)(d__component)
