@@ -1,10 +1,12 @@
 
-const p__immutable = require('immutable')
+import * as p__immutable from 'immutable'
 
-const m__action_types = require('./action-types')
-const m__redux_helpers = require('~/redux-helpers')
-const m__state = require('./state')
-const m__token = require('./token')
+import * as m__action_types from './action-types'
+import * as m__redux_helpers from '~/redux-helpers'
+import * as m__token from './token'
+
+import d__state from './state'
+
 
 const reducers = {
 	[m__action_types.create]: (state, action) => {
@@ -16,7 +18,7 @@ const reducers = {
 		})
 		return state.insert(0, value)
 	},
-	[m__action_types.delete]: (state, action) => {
+	[m__action_types.remove]: (state, action) => {
 		const index = state.findIndex((value) => {
 			return value.get('id') === action.payload.id
 		})
@@ -48,8 +50,8 @@ const reducers = {
 	},
 }
 
-module.exports = m__redux_helpers.reducers.middlewares_composer([
-	m__redux_helpers.middlewares.state_initializer(m__state),
+export default m__redux_helpers.reducers.middlewares_composer([
+	m__redux_helpers.middlewares.state_initializer(d__state),
 	m__redux_helpers.middlewares.filters_applicator([
 		m__redux_helpers.reducers.tokens_checker(m__token.array),
 	]),
