@@ -5,10 +5,10 @@ module.exports = {
 	],
 	plugins: [
 		'@babel/plugin-proposal-class-properties',
+		'@babel/plugin-proposal-export-namespace-from',
 		'@babel/plugin-syntax-dynamic-import',
 		'@babel/plugin-transform-runtime',
 		'babel-plugin-lodash',
 		'react-hot-loader/babel',
 	],
-	sourceType: 'script',
 }
