@@ -1,5 +1,5 @@
 
-exports.styler = (styles) => {
+export const styler = (styles) => {
 	return (...keys) => {
 		return ['', ...keys].map((key) => {
 			return styles[key]
@@ -7,7 +7,7 @@ exports.styler = (styles) => {
 	}
 }
 
-exports.cast_actions = (token, array) => {
+export const cast_actions = (token, array) => {
 	return array.reduce((object, value) => {
 		object[value] = [token, value].join('--')
 		return object
